chore 📦: Update CI workflow to include versioning and package install… (#1)
* chore 📦: Update CI workflow to include versioning and package installation

Updated the .github/workflows/ci.yml file to install dependencies, run Ruff, mypy, and pytest on every push and pull request. Also added documentation for CI & Versioning.

* moved build files to src directory.

* refactor ♻️: Update media assets and audio files in demo article

Updated the media assets and audio files in the demo article to reflect the new design and layout, ensuring consistency with the rest of the content. This change also includes updating the audio file paths to match the new directory structure.

* chore 📦: Update dependencies and scripts

Updated the project's dependencies to include new libraries for Hugging Face and PyTorch, and updated the scripts section in pyproject.toml to reflect these changes.

* style 💎: Update template rendering to check for string output

Fixed a bug where the `render` method of Jinja2 templates could return non-string values, causing a ThemeError. This change ensures that the rendered template is always a string, preventing potential errors.