Watermarking (#18)
* style 💎: Improve tag display in journal cards

Updated the `renderTags` function to limit the number of tags displayed per card, and added a 'more' link for hidden tags. Also updated the template for tile-article.html to include the data-tile-tags attribute.

* docs 📝: Update media processing documentation with new watermark and metadata embedding features

This commit updates the documentation for media processing, including the addition of a new watermark feature and improved metadata embedding. The changes provide more detailed information on configuring the watermark overlay and embedding copyright/licensing metadata into output files.

* docs 📝: Update Inter OFL License to version 1.1

Added the SIL Open Font License (OFL) Version 1.1 to the documentation, providing a free and open framework for font sharing and improvement.

* style 💎: Update type hint for media processor metadata

Updated the `MediaMetadataEmbedConfig` type to include a generic `Any` type for its dictionary values, allowing for more flexibility in embedding metadata. This change improves code readability and maintainability.

* style 💎: Ensure consistent type across code paths for mypy

Consistently typed the `font` variable to avoid potential type errors when using mypy. This change ensures that the font is loaded correctly regardless of whether a custom path is provided or not.

* refactor ♻️: Improved type hinting for font usage

Updated type hints for `font` variable to include Union type, ensuring consistency across code paths and improving mypy compatibility.